Rename error.py files to exceptions.py
We've discussed this in the past: Pyramid exception views are called
exceptions views not just error views, and can catch any kind of
exception and return any kind of response, not just errors.

Just implementing the name change now.
