+++ lms/views/api/exceptions.py
